% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aseg_all}
\alias{aseg_all}
\alias{aseg_cerebrum}
\title{ggseg tissue class atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.
}
\usage{
aseg_all

aseg_cerebrum
}
\description{
2D ggseg atlas for plotting based on brain tissue type. Atlas is based on the FreeSurfer aseg atlas segmentations and was created with data-raw/aseg_all_atlas.Rmd. Atlas parses the brain into brainstem, cerebellum, cortical GM, CSF, subcortical GM, and WM. Adjacent polygons within each tissue class are merged to create continuous regions.  
Version 1.1; updated with polygon merging.
}
\examples{
library(ggsegTissue)
library(ggplot2)
#make dataframe
df <- data.frame(my_variable = c(1, 3, 7, 4),
                 tissue_class = c("cGM", "CSF", "sGM", "WM"))

#plot
ggplot(df) +
  ggseg::geom_brain(atlas = aseg_all, 
             aes(fill = my_variable)) +
             ggtitle("This is my brain plot!")
}
\references{
Fischl et al., (2002). Neuron, 33:341-355 (\href{https://www.pubmed.ncbi.nlm.nih.gov/11832223}{PubMed})

\itemize{
 \item{aseg_all}{ - aseg brain tissue class atlas}
 \item{aseg_cerebrum}{ - aseg brain tissue class atlas without cerebellum or brainstem}
}
}
\concept{ggseg_atlases}
\keyword{datasets}
